Bug fixes and code cleanup for IRequests without a response - Refactored wrappers into individual classes - Fixed the request handler wrapper with no response to not require Unit
