+++ src/Flume/Mediator.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Flume.Internal;
 using Flume.NotificationPublishers;
+using Flume.Wrappers;
 
 namespace Flume;
